begin transform refactor - split off utils and biodomains
